+++ src/masonite/providers/AuthenticationProvider.py
@@ -14,4 +14,5 @@ class AuthenticationProvider(Provider):
         self.application.bind("auth", auth)
 
     def boot(self):
-        self.application.make("auth").guard("web").logout()
+        # self.application.make("auth").guard("web").logout()
+        pass
